printf() format warning fixed
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@18244 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/generic/grid.cpp
@@ -967,7 +967,7 @@ wxString wxGridCellNumberEditor::GetValue() const
 
   if( HasRange() )
   {
-    int value = Spin()->GetValue();
+    long value = Spin()->GetValue();
     s.Printf(wxT("%ld"), value);
   }
   else
